feat: enhance classroom and student sections with initial data support
- Updated ClassroomSection to accept initial classrooms data, reducing loading time.
- Modified StudentSection to handle initial batches, improving user experience on load.
- Implemented error handling and snapshot management in AssignmentWorkspace for better reliability.
- Introduced periodic snapshot flushing to preserve user progress during assignments.
- Added new API endpoints for managing snapshots and linking them to submissions.
- Enhanced API configuration for better environment handling and debugging.
- Updated database schema to support new snapshot functionality and replay logs.

+++ gyaansetu-api/app/models/submission.py
@@ -28,7 +28,7 @@ class BehaviorLogCreate(BaseModel):
     paste_events: list[dict]
     largest_paste: int
     tab_switches: int
-    idle_time: float
+    idle_time: int
 
 
 class UploadCreate(BaseModel):
